# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/budgets/{budgetId}:
    get:
      tags:
      - Budgets
      summary: Get budget by ID
      description: Returns the specified budget belonging to the authenticated user.
      operationId: getBudgetById
      parameters:
      - name: budgetId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Budget retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        "404":
          description: Budget not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
    put:
      tags:
      - Budgets
      summary: Update a budget
      description: Updates the specified budget belonging to the authenticated user
        with the provided data.
      operationId: updateBudget
      parameters:
      - name: budgetId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetRequest"
        required: true
      responses:
        "200":
          description: Budget updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        "400":
          description: Invalid budget data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        "404":
          description: Budget not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        "409":
          description: A different budget already exists for this month
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
    delete:
      tags:
      - Budgets
      summary: Delete a budget
      description: Deletes the specified budget belonging to the authenticated user.
      operationId: deleteBudget
      parameters:
      - name: budgetId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Deletes the specified budget belonging to the authenticated
            user
        "404":
          description: Budget not found
...
  /api/budgets:
    get:
      tags:
      - Budgets
      summary: Get budget by period
      description: Returns the authenticated user's budget for the specified budget
        period.
      operationId: getBudgetByPeriod
      parameters:
      - name: period
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: Budget retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        "404":
          description: Budget not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
    post:
      tags:
      - Budgets
      summary: Create a budget
      description: Creates a new budget for the authenticated user for the specified
        budget period.
      operationId: createBudget
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetRequest"
        required: true
      responses:
        "201":
          description: Budget created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        "400":
          description: Invalid budget data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
        "409":
          description: A budget already exists for the specified period
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register a new user
      description: "Creates a new user account using the provided username, email,\
        \ and password."
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: Registration successful
        "409":
          description: Username or Email already used
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: Log in a user
      description: Authenticates a user using their username and password and returns
        a JWT token.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Authentication successful
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid username or password
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /api/budgets/summary:
    get:
      tags:
      - Budgets
      summary: Get budget summary
      description: Returns a summary of the authenticated user's budget and expenses
        for the specified budget period.
      operationId: getBudgetSummary
      parameters:
      - name: period
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: Budget summary retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetSummaryResponse"
        "404":
          description: Budget not found for the specified period
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetSummaryResponse"
components:
  schemas:
...
    BudgetRequest:
      type: object
      properties:
        amount:
          type: number
          exclusiveMinimum: 0
        budgetPeriod:
          type: string
          format: date
      required:
      - amount
      - budgetPeriod
    BudgetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        amount:
          type: number
        budgetPeriod:
          type: string
          format: date
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 0
        firstName:
          type: string
          maxLength: 50
          minLength: 0
        lastName:
          type: string
          maxLength: 50
          minLength: 0
        password:
          type: string
          maxLength: 255
          minLength: 0
        email:
          type: string
          maxLength: 100
          minLength: 0
      required:
      - email
      - firstName
      - lastName
      - password
      - username
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 0
        password:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - password
      - username
    LoginResponse:
      type: object
      properties:
        token:
          type: string
...
    BudgetSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        amount:
          type: number
        budgetPeriod:
          type: string
          format: date
        totalCurrentExpenses:
          type: number
        totalMonthlyExpenses:
          type: number
        remainingCurrentBudget:
          type: number
        remainingMonthlyBudget:
          type: number
        budgetPercentage:
          type: number